test: merge ang. dist. conftest into the tests
