--- tests/physics/helicity_formalism/angular_distributions/__init__.py
@@ -1,1 +1,0 @@
-"""Required to set mypy options for the tests folder."""
